% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{create_change_set}
\alias{create_change_set}
\title{Creates a list of changes that will be applied to a stack so that you can review the changes before executing them}
\usage{
create_change_set(StackName, ChangeSetName, TemplateBody = NULL,
  TemplateURL = NULL, UsePreviousTemplate = NULL, Parameters = NULL,
  Capabilities = NULL, ResourceTypes = NULL, RoleARN = NULL,
  RollbackConfiguration = NULL, NotificationARNs = NULL, Tags = NULL,
  ClientToken = NULL, Description = NULL, ChangeSetType = NULL)
}
\arguments{
\item{StackName}{The name or the unique ID of the stack for which you are creating a change set. AWS CloudFormation generates the change set by comparing this stack's information with the information that you submit, such as a modified template or different parameter input values.}

\item{ChangeSetName}{The name of the change set. The name must be unique among all change sets that are associated with the specified stack.

A change set name can contain only alphanumeric, case sensitive characters and hyphens. It must start with an alphabetic character and cannot exceed 128 characters.}

\item{TemplateBody}{A structure that contains the body of the revised template, with a minimum length of 1 byte and a maximum length of 51,200 bytes. AWS CloudFormation generates the change set by comparing this template with the template of the stack that you specified.

Conditional: You must specify only \code{TemplateBody} or \code{TemplateURL}.}

\item{TemplateURL}{The location of the file that contains the revised template. The URL must point to a template (max size: 460,800 bytes) that is located in an S3 bucket. AWS CloudFormation generates the change set by comparing this template with the stack that you specified.

Conditional: You must specify only \code{TemplateBody} or \code{TemplateURL}.}

\item{UsePreviousTemplate}{Whether to reuse the template that is associated with the stack to create the change set.}

\item{Parameters}{A list of \code{Parameter} structures that specify input parameters for the change set. For more information, see the \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_Parameter.html}{Parameter} data type.}

\item{Capabilities}{A list of values that you must specify before AWS CloudFormation can update certain stacks. Some stack templates might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those stacks, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are \code{CAPABILITY_IAM} and \code{CAPABILITY_NAMED_IAM}. The following resources require you to specify this parameter: \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-accesskey.html}{AWS::IAM::AccessKey}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-group.html}{AWS::IAM::Group}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-iam-instanceprofile.html}{AWS::IAM::InstanceProfile}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-policy.html}{AWS::IAM::Policy}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-iam-role.html}{AWS::IAM::Role}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-user.html}{AWS::IAM::User}, and \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-addusertogroup.html}{AWS::IAM::UserToGrou

If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify \code{CAPABILITY_NAMED_IAM}. If you don't specify this parameter, this action returns an \code{InsufficientCapabilities} error.

For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-iam-template.html#capabilities}{Acknowledging IAM Resources in AWS CloudFormation Templates}.}

\item{ResourceTypes}{The template resource types that you have permissions to work with if you execute this change set, such as \code{AWS::EC2::Instance}, \code{AWS::EC2::*}, or \code{Custom::MyCustomInstance}.

If the list of resource types doesn't include a resource type that you're updating, the stack update fails. By default, AWS CloudFormation grants permissions to all resource types. AWS Identity and Access Management (IAM) uses this parameter for condition keys in IAM policies for AWS CloudFormation. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-iam-template.html}{Controlling Access with AWS Identity and Access Management} in the AWS CloudFormation User Guide.}

\item{RoleARN}{The Amazon Resource Name (ARN) of an AWS Identity and Access Management (IAM) role that AWS CloudFormation assumes when executing the change set. AWS CloudFormation uses the role's credentials to make calls on your behalf. AWS CloudFormation uses this role for all future operations on the stack. As long as users have permission to operate on the stack, AWS CloudFormation uses this role even if the users don't have permission to pass it. Ensure that the role grants least privilege.

If you don't specify a value, AWS CloudFormation uses the role that was previously associated with the stack. If no role is available, AWS CloudFormation uses a temporary session that is generated from your user credentials.}

\item{RollbackConfiguration}{The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.}

\item{NotificationARNs}{The Amazon Resource Names (ARNs) of Amazon Simple Notification Service (Amazon SNS) topics that AWS CloudFormation associates with the stack. To remove all associated notification topics, specify an empty list.}

\item{Tags}{Key-value pairs to associate with this stack. AWS CloudFormation also propagates these tags to resources in the stack. You can specify a maximum of 50 tags.}

\item{ClientToken}{A unique identifier for this \code{CreateChangeSet} request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another change set with the same name. You might retry \code{CreateChangeSet} requests to ensure that AWS CloudFormation successfully received them.}

\item{Description}{A description to help you identify this change set.}

\item{ChangeSetType}{The type of change set operation. To create a change set for a new stack, specify \code{CREATE}. To create a change set for an existing stack, specify \code{UPDATE}.

If you create a change set for a new stack, AWS Cloudformation creates a stack with a unique stack ID, but no template or resources. The stack will be in the \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-describing-stacks.html#d0e11995}{REVIEW_IN_PROGRESS} state until you execute the change set.

By default, AWS CloudFormation specifies \code{UPDATE}. You can't use the \code{UPDATE} type to create a change set for a new stack or the \code{CREATE} type to create a change set for an existing stack.}
}
\description{
Creates a list of changes that will be applied to a stack so that you can review the changes before executing them. You can create a change set for a stack that doesn't exist or an existing stack. If you create a change set for a stack that doesn't exist, the change set shows all of the resources that AWS CloudFormation will create. If you create a change set for an existing stack, AWS CloudFormation compares the stack's information with the information that you submit in the change set and lists the differences. Use change sets to understand which resources AWS CloudFormation will create or change, and how it will change resources in an existing stack, before you create or update a stack.
}
\details{
To create a change set for a stack that doesn't exist, for the \code{ChangeSetType} parameter, specify \code{CREATE}. To create a change set for an existing stack, specify \code{UPDATE} for the \code{ChangeSetType} parameter. After the \code{CreateChangeSet} call successfully completes, AWS CloudFormation starts creating the change set. To check the status of the change set or to review it, use the DescribeChangeSet action.

When you are satisfied with the changes the change set will make, execute the change set by using the ExecuteChangeSet action. AWS CloudFormation doesn't make changes until you execute the change set.
}
\examples{

}
