% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{create_stack_set}
\alias{create_stack_set}
\title{Creates a stack set}
\usage{
create_stack_set(StackSetName, Description = NULL, TemplateBody = NULL,
  TemplateURL = NULL, Parameters = NULL, Capabilities = NULL,
  Tags = NULL, AdministrationRoleARN = NULL, ExecutionRoleName = NULL,
  ClientRequestToken = NULL)
}
\arguments{
\item{StackSetName}{The name to associate with the stack set. The name must be unique in the region where you create your stack set.

A stack name can contain only alphanumeric characters (case-sensitive) and hyphens. It must start with an alphabetic character and can't be longer than 128 characters.}

\item{Description}{A description of the stack set. You can use the description to identify the stack set's purpose or other important information.}

\item{TemplateBody}{The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/template-anatomy.html}{Template Anatomy} in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.}

\item{TemplateURL}{The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that's located in an Amazon S3 bucket. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/template-anatomy.html}{Template Anatomy} in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.}

\item{Parameters}{The input parameters for the stack set template.}

\item{Capabilities}{A list of values that you must specify before AWS CloudFormation can create certain stack sets. Some stack set templates might include resources that can affect permissions in your AWS account---for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY\_IAM and CAPABILITY\_NAMED\_IAM. The following resources require you to specify this parameter:
\itemize{
\item AWS::IAM::AccessKey
\item AWS::IAM::Group
\item AWS::IAM::InstanceProfile
\item AWS::IAM::Policy
\item AWS::IAM::Role
\item AWS::IAM::User
\item AWS::IAM::UserToGroupAddition
}

If your stack template contains these resources, we recommend that you review all permissions that are associated with them and edit their permissions if necessary.

If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify CAPABILITY\_NAMED\_IAM. If you don't specify this parameter, this action returns an \code{InsufficientCapabilities} error.

For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-iam-template.html#capabilities}{Acknowledging IAM Resources in AWS CloudFormation Templates.}}

\item{Tags}{The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. A maximum number of 50 tags can be specified.

If you specify tags as part of a \code{CreateStackSet} action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you don't, the entire \code{CreateStackSet} action fails with an \code{access denied} error, and the stack set is not created.}

\item{AdministrationRoleARN}{The Amazon Resource Number (ARN) of the IAM role to use to create this stack set.

Specify an IAM role only if you are using customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/stacksets-prereqs.html}{Prerequisites: Granting Permissions for Stack Set Operations} in the \emph{AWS CloudFormation User Guide}.}

\item{ExecutionRoleName}{The name of the IAM execution role to use to create the stack set. If you do not specify an execution role, AWS CloudFormation uses the \code{AWSCloudFormationStackSetExecutionRole} role for the stack set operation.

Specify an IAM role only if you are using customized execution roles to control which stack resources users and groups can include in their stack sets.}

\item{ClientRequestToken}{A unique identifier for this \code{CreateStackSet} request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another stack set with the same name. You might retry \code{CreateStackSet} requests to ensure that AWS CloudFormation successfully received them.

If you don't specify an operation ID, the SDK generates one automatically.}
}
\description{
Creates a stack set.
}
\examples{

}
