% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{delete_stack_set}
\alias{delete_stack_set}
\title{Deletes a stack set}
\usage{
delete_stack_set(StackSetName)
}
\arguments{
\item{StackSetName}{The name or unique ID of the stack set that you're deleting. You can obtain this value by running ListStackSets.}
}
\description{
Deletes a stack set. Before you can delete a stack set, all of its member stack instances must be deleted. For more information about how to do this, see DeleteStackInstances.
}
\examples{

}
