% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{describe_change_set}
\alias{describe_change_set}
\title{Returns the inputs for the change set and a list of changes that AWS CloudFormation will make if you execute the change set}
\usage{
describe_change_set(ChangeSetName, StackName = NULL, NextToken = NULL)
}
\arguments{
\item{ChangeSetName}{The name or Amazon Resource Name (ARN) of the change set that you want to describe.}

\item{StackName}{If you specified the name of a change set, specify the stack name or ID (ARN) of the change set you want to describe.}

\item{NextToken}{A string (provided by the DescribeChangeSet response output) that identifies the next page of information that you want to retrieve.}
}
\description{
Returns the inputs for the change set and a list of changes that AWS CloudFormation will make if you execute the change set. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-changesets.html}{Updating Stacks Using Change Sets} in the AWS CloudFormation User Guide.
}
\examples{

}
