% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{describe_stack_instance}
\alias{describe_stack_instance}
\title{Returns the stack instance that's associated with the specified stack set, AWS account, and region}
\usage{
describe_stack_instance(
  StackSetName = "string",
  StackInstanceAccount = "string",
  StackInstanceRegion = "string"
)
}
\arguments{
\item{StackSetName}{[required] The name or the unique stack ID of the stack set that you want to get stack instance information for.}

\item{StackInstanceAccount}{[required] The ID of an AWS account that's associated with this stack instance.}

\item{StackInstanceRegion}{[required] The name of a region that's associated with this stack instance.}
}
\description{
Returns the stack instance that's associated with the specified stack set, AWS account, and region.
}
\details{
For a list of stack instances that are associated with a specific stack set, use ListStackInstances.
}
\examples{

}
