% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{describe_stack_set}
\alias{describe_stack_set}
\title{Returns the description of the specified stack set}
\usage{
describe_stack_set(
  StackSetName = "string"
)
}
\arguments{
\item{StackSetName}{[required] The name or unique ID of the stack set whose description you want.}
}
\description{
Returns the description of the specified stack set.
}
\examples{

}
