% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{describe_stack_set_operation}
\alias{describe_stack_set_operation}
\title{Returns the description of the specified stack set operation}
\usage{
describe_stack_set_operation(
  StackSetName = "string",
  OperationId = "string"
)
}
\arguments{
\item{StackSetName}{[required] The name or the unique stack ID of the stack set for the stack operation.}

\item{OperationId}{[required] The unique ID of the stack set operation.}
}
\description{
Returns the description of the specified stack set operation.
}
\examples{

}
