% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{get_stack_policy}
\alias{get_stack_policy}
\title{Returns the stack policy for a specified stack}
\usage{
get_stack_policy(StackName)
}
\arguments{
\item{StackName}{The name or unique stack ID that is associated with the stack whose policy you want to get.}
}
\description{
Returns the stack policy for a specified stack. If a stack doesn't have a policy, a null value is returned.
}
\examples{

}
