% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{list_change_sets}
\alias{list_change_sets}
\title{Returns the ID and status of each active change set for a stack}
\usage{
list_change_sets(StackName, NextToken = NULL)
}
\arguments{
\item{StackName}{The name or the Amazon Resource Name (ARN) of the stack for which you want to list change sets.}

\item{NextToken}{A string (provided by the ListChangeSets response output) that identifies the next page of change sets that you want to retrieve.}
}
\description{
Returns the ID and status of each active change set for a stack. For example, AWS CloudFormation lists change sets that are in the \code{CREATE_IN_PROGRESS} or \code{CREATE_PENDING} state.
}
\examples{

}
