% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{list_exports}
\alias{list_exports}
\title{Lists all exported output values in the account and region in which you call this action}
\usage{
list_exports(NextToken = NULL)
}
\arguments{
\item{NextToken}{A string (provided by the ListExports response output) that identifies the next page of exported output values that you asked to retrieve.}
}
\description{
Lists all exported output values in the account and region in which you call this action. Use this action to see the exported output values that you can import into other stacks. To import values, use the \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-importvalue.html}{Fn::ImportValue} function.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-stack-exports.html}{AWS CloudFormation Export Stack Output Values}.
}
\examples{

}
