% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{list_imports}
\alias{list_imports}
\title{Lists all stacks that are importing an exported output value}
\usage{
list_imports(
  ExportName = "string",
  NextToken = "string"
)
}
\arguments{
\item{ExportName}{[required] The name of the exported output value. AWS CloudFormation returns the stack names that are importing this value.}

\item{NextToken}{A string (provided by the ListImports response output) that identifies the next page of stacks that are importing the specified exported output value.}
}
\description{
Lists all stacks that are importing an exported output value. To modify or remove an exported output value, first use this action to see which stacks are using it. To see the exported output values in your account, see ListExports.
}
\details{
For more information about importing an exported output value, see the \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-importvalue.html}{Fn::ImportValue} function.
}
\examples{

}
