% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{set_stack_policy}
\alias{set_stack_policy}
\title{Sets a stack policy for a specified stack}
\usage{
set_stack_policy(StackName, StackPolicyBody = NULL,
  StackPolicyURL = NULL)
}
\arguments{
\item{StackName}{[required] The name or unique stack ID that you want to associate a policy with.}

\item{StackPolicyBody}{Structure containing the stack policy body. For more information, go to \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/protect-stack-resources.html}{Prevent Updates to Stack Resources} in the AWS CloudFormation User Guide. You can specify either the \code{StackPolicyBody} or the \code{StackPolicyURL} parameter, but not both.}

\item{StackPolicyURL}{Location of a file containing the stack policy. The URL must point to a policy (maximum size: 16 KB) located in an S3 bucket in the same region as the stack. You can specify either the \code{StackPolicyBody} or the \code{StackPolicyURL} parameter, but not both.}
}
\description{
Sets a stack policy for a specified stack.
}
\section{Accepted Parameters}{
\preformatted{set_stack_policy(
  StackName = "string",
  StackPolicyBody = "string",
  StackPolicyURL = "string"
)
}
}

