% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{update_stack}
\alias{update_stack}
\title{Updates a stack as specified in the template}
\usage{
update_stack(StackName, TemplateBody = NULL, TemplateURL = NULL,
  UsePreviousTemplate = NULL, StackPolicyDuringUpdateBody = NULL,
  StackPolicyDuringUpdateURL = NULL, Parameters = NULL,
  Capabilities = NULL, ResourceTypes = NULL, RoleARN = NULL,
  RollbackConfiguration = NULL, StackPolicyBody = NULL,
  StackPolicyURL = NULL, NotificationARNs = NULL, Tags = NULL,
  ClientRequestToken = NULL)
}
\arguments{
\item{StackName}{[required] The name or unique stack ID of the stack to update.}

\item{TemplateBody}{Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/template-anatomy.html}{Template Anatomy} in the AWS CloudFormation User Guide.)

Conditional: You must specify only one of the following parameters: \code{TemplateBody}, \code{TemplateURL}, or set the \code{UsePreviousTemplate} to \code{true}.}

\item{TemplateURL}{Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket. For more information, go to \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/template-anatomy.html}{Template Anatomy} in the AWS CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: \code{TemplateBody}, \code{TemplateURL}, or set the \code{UsePreviousTemplate} to \code{true}.}

\item{UsePreviousTemplate}{Reuse the existing template that is associated with the stack that you are updating.

Conditional: You must specify only one of the following parameters: \code{TemplateBody}, \code{TemplateURL}, or set the \code{UsePreviousTemplate} to \code{true}.}

\item{StackPolicyDuringUpdateBody}{Structure containing the temporary overriding stack policy body. You can specify either the \code{StackPolicyDuringUpdateBody} or the \code{StackPolicyDuringUpdateURL} parameter, but not both.

If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.}

\item{StackPolicyDuringUpdateURL}{Location of a file containing the temporary overriding stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the \code{StackPolicyDuringUpdateBody} or the \code{StackPolicyDuringUpdateURL} parameter, but not both.

If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.}

\item{Parameters}{A list of \code{Parameter} structures that specify input parameters for the stack. For more information, see the \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_Parameter.html}{Parameter} data type.}

\item{Capabilities}{A list of values that you must specify before AWS CloudFormation can update certain stacks. Some stack templates might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those stacks, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are \code{CAPABILITY_IAM} and \code{CAPABILITY_NAMED_IAM}. The following resources require you to specify this parameter: \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-accesskey.html}{AWS::IAM::AccessKey}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-group.html}{AWS::IAM::Group}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-iam-instanceprofile.html}{AWS::IAM::InstanceProfile}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-policy.html}{AWS::IAM::Policy}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-iam-role.html}{AWS::IAM::Role}, \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-user.html}{AWS::IAM::User}, and \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-addusertogroup.html}{AWS::IAM::UserToGrou

If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify \code{CAPABILITY_NAMED_IAM}. If you don't specify this parameter, this action returns an \code{InsufficientCapabilities} error.

For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-iam-template.html#capabilities}{Acknowledging IAM Resources in AWS CloudFormation Templates}.}

\item{ResourceTypes}{The template resource types that you have permissions to work with for this update stack action, such as \code{AWS::EC2::Instance}, \code{AWS::EC2::*}, or \code{Custom::MyCustomInstance}.

If the list of resource types doesn't include a resource that you're updating, the stack update fails. By default, AWS CloudFormation grants permissions to all resource types. AWS Identity and Access Management (IAM) uses this parameter for AWS CloudFormation-specific condition keys in IAM policies. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-iam-template.html}{Controlling Access with AWS Identity and Access Management}.}

\item{RoleARN}{The Amazon Resource Name (ARN) of an AWS Identity and Access Management (IAM) role that AWS CloudFormation assumes to update the stack. AWS CloudFormation uses the role's credentials to make calls on your behalf. AWS CloudFormation always uses this role for all future operations on the stack. As long as users have permission to operate on the stack, AWS CloudFormation uses this role even if the users don't have permission to pass it. Ensure that the role grants least privilege.

If you don't specify a value, AWS CloudFormation uses the role that was previously associated with the stack. If no role is available, AWS CloudFormation uses a temporary session that is generated from your user credentials.}

\item{RollbackConfiguration}{The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.}

\item{StackPolicyBody}{Structure containing a new stack policy body. You can specify either the \code{StackPolicyBody} or the \code{StackPolicyURL} parameter, but not both.

You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.}

\item{StackPolicyURL}{Location of a file containing the updated stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the \code{StackPolicyBody} or the \code{StackPolicyURL} parameter, but not both.

You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.}

\item{NotificationARNs}{Amazon Simple Notification Service topic Amazon Resource Names (ARNs) that AWS CloudFormation associates with the stack. Specify an empty list to remove all notification topics.}

\item{Tags}{Key-value pairs to associate with this stack. AWS CloudFormation also propagates these tags to supported resources in the stack. You can specify a maximum number of 50 tags.

If you don't specify this parameter, AWS CloudFormation doesn't modify the stack's tags. If you specify an empty value, AWS CloudFormation removes all associated tags.}

\item{ClientRequestToken}{A unique identifier for this \code{UpdateStack} request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to update a stack with the same name. You might retry \code{UpdateStack} requests to ensure that AWS CloudFormation successfully received them.

All events triggered by a given stack operation are assigned the same client request token, which you can use to track operations. For example, if you execute a \code{CreateStack} operation with the token \code{token1}, then all the \code{StackEvents} generated by that operation will have \code{ClientRequestToken} set as \code{token1}.

In the console, stack operations display the client request token on the Events tab. Stack operations that are initiated from the console use the token format \emph{Console-StackOperation-ID}, which helps you easily identify the stack operation . For example, if you create a stack using the console, each stack event would be assigned the same token in the following format: \code{Console-CreateStack-7f59c3cf-00d2-40c7-b2ff-e75db0987002}.}
}
\description{
Updates a stack as specified in the template. After the call completes successfully, the stack update starts. You can check the status of the stack via the DescribeStacks action.
}
\details{
To get a copy of the template for an existing stack, you can use the GetTemplate action.

For more information about creating an update template, updating a stack, and monitoring the progress of the update, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks.html}{Updating a Stack}.
}
\section{Accepted Parameters}{
\preformatted{update_stack(
  StackName = "string",
  TemplateBody = "string",
  TemplateURL = "string",
  UsePreviousTemplate = TRUE|FALSE,
  StackPolicyDuringUpdateBody = "string",
  StackPolicyDuringUpdateURL = "string",
  Parameters = list(
    list(
      ParameterKey = "string",
      ParameterValue = "string",
      UsePreviousValue = TRUE|FALSE,
      ResolvedValue = "string"
    )
  ),
  Capabilities = list(
    "CAPABILITY_IAM"|"CAPABILITY_NAMED_IAM"
  ),
  ResourceTypes = list(
    "string"
  ),
  RoleARN = "string",
  RollbackConfiguration = list(
    RollbackTriggers = list(
      list(
        Arn = "string",
        Type = "string"
      )
    ),
    MonitoringTimeInMinutes = 123
  ),
  StackPolicyBody = "string",
  StackPolicyURL = "string",
  NotificationARNs = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  ClientRequestToken = "string"
)
}
}

