% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudformation_operations.R
\name{update_termination_protection}
\alias{update_termination_protection}
\title{Updates termination protection for the specified stack}
\usage{
update_termination_protection(EnableTerminationProtection, StackName)
}
\arguments{
\item{EnableTerminationProtection}{Whether to enable termination protection on the specified stack.}

\item{StackName}{The name or unique ID of the stack for which you want to set termination protection.}
}
\description{
Updates termination protection for the specified stack. If a user attempts to delete a stack with termination protection enabled, the operation fails and the stack remains unchanged. For more information, see \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-protect-stacks.html}{Protecting a Stack From Being Deleted} in the \emph{AWS CloudFormation User Guide}.
}
\details{
For \href{http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-nested-stacks.html}{nested stacks}, termination protection is set on the root stack and cannot be changed directly on the nested stack.
}
\examples{

}
